server/requirements.txt
Flask
openai-whisper
torch
torchaudio